Installed supabase deps and moved sanity lib files into a specific directory
