+++ apps/www/package.json
@@ -31,6 +31,7 @@
 		"@sanity/asset-utils": "^1.3.0",
 		"@sanity/client": "^6.7.1",
 		"@sanity/image-url": "^1.0.2",
+		"@supabase/ssr": "^0.1.0",
 		"@supabase/supabase-js": "^2.38.4",
 		"@vercel/og": "^0.5.20",
 		"boring-avatars": "^1.10.1",
