+++ apps/www/src/app/(pages)/articles/page.tsx
@@ -1,4 +1,4 @@
-import { sanityClient } from '@lib/sanity.client'
+import { sanityClient } from '@lib/sanity/sanity.client'
 import { articlesPageQuery } from '@lib/queries'
 import { getOgTitle } from '@utils/metaHelpers'
