+++ apps/www/src/app/(pages)/contributors/[slug]/page.tsx
@@ -1,7 +1,7 @@
 import { notFound } from 'next/navigation'
 import { draftMode } from 'next/headers'
-import { readToken } from '@lib/sanity.api'
-import { sanityClient } from '@lib/sanity.client'
+import { readToken } from '@lib/sanity/sanity.api'
+import { sanityClient } from '@lib/sanity/sanity.client'
 import { contributorsBySlugQuery, contributorPaths } from '@lib/queries'
 import { getSanityImageUrl } from '@utils/getSanityImage'
 import { getOgTitle } from '@utils/metaHelpers'
